+++ bootique/src/test/java/io/bootique/value/BytesTest.java
@@ -1,0 +1,126 @@
+package io.bootique.value;
+
+import org.junit.Test;
+
+import static io.bootique.value.BytesUnit.*;
+import static org.junit.Assert.*;
+
+public class BytesTest {
+
+    @Test
+    public void testParse_byte() {
+        String checkString = 5 + " " + BYTES.getName();
+        assertEquals(checkString, new Bytes("5b").toString());
+        assertEquals(checkString, new Bytes("5 b").toString());
+        assertEquals(checkString, new Bytes("5byte").toString());
+        assertEquals(checkString, new Bytes("5 byte").toString());
+        assertEquals(checkString, new Bytes("5bytes").toString());
+        assertEquals(checkString, new Bytes("5 bytes").toString());
+    }
+
+    @Test
+    public void testParse_KB() {
+        assertEquals(5 + " " + KB.getName(),  new Bytes("5kb").toString());
+        assertEquals(5120, new Bytes("5kb").getBytes());
+        assertEquals(5120, new Bytes("5 kb").getBytes());
+        assertEquals(5120, new Bytes("5kilobyte").getBytes());
+        assertEquals(5120, new Bytes("5 kilobyte").getBytes());
+        assertEquals(5120, new Bytes("5kilobytes").getBytes());
+        assertEquals(5120, new Bytes("5 kilobytes").getBytes());
+    }
+
+    @Test
+    public void testParse_MB() {
+        assertEquals(5 + " " + MB.getName(),  new Bytes("5mb").toString());
+        assertEquals(5242880, new Bytes("5mb").getBytes());
+        assertEquals(5242880, new Bytes("5 mb").getBytes());
+        assertEquals(5242880, new Bytes("5megabyte").getBytes());
+        assertEquals(5242880, new Bytes("5 megabyte").getBytes());
+        assertEquals(5242880, new Bytes("5megabytes").getBytes());
+        assertEquals(5242880, new Bytes("5 megabytes").getBytes());
+    }
+
+    @Test
+    public void testParse_GB() {
+        assertEquals(5 + " " + GB.getName(),  new Bytes("5gb").toString());
+        assertEquals(5368709120L, new Bytes("5gb").getBytes());
+        assertEquals(5368709120L, new Bytes("5 gb").getBytes());
+        assertEquals(5368709120L, new Bytes("5gigabyte").getBytes());
+        assertEquals(5368709120L, new Bytes("5 gigabyte").getBytes());
+        assertEquals(5368709120L, new Bytes("5gigabytes").getBytes());
+        assertEquals(5368709120L, new Bytes("5 gigabytes").getBytes());
+    }
+
+    @Test(expected = NullPointerException.class)
+    public void testParse_Null() {
+        new Bytes(null);
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testParse_Empty() {
+        new Bytes("");
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testParse_Invalid1() {
+        new Bytes("4 nosuchthing");
+    }
+
+    @Test(expected = IllegalArgumentException.class)
+    public void testParse_Invalid2() {
+        new Bytes("not_a_number sec");
+    }
+
+
+    @Test
+    public void testCompareTo() {
+        Bytes b1 = new Bytes("1b");
+        Bytes b2 = new Bytes("2b");
+        Bytes b3 = new Bytes("2 bytes");
+        Bytes b4 = new Bytes("2kb");
+        Bytes b5 = new Bytes("2mb");
+        Bytes b6 = new Bytes("2gb");
+
+        assertTrue(b1.compareTo(b1) == 0);
+        assertTrue(b1.compareTo(b2) < 0);
+        assertTrue(b2.compareTo(b1) > 0);
+        assertTrue(b2.compareTo(b3) == 0);
+        assertTrue(b4.compareTo(b2) > 0);
+        assertTrue(b5.compareTo(b4) > 0);
+        assertTrue(b6.compareTo(b5) > 0);
+
+    }
+
+    @Test
+    public void testEquals() {
+        Bytes b1 = new Bytes("5kb");
+        Bytes b2 = new Bytes("5368709120b");
+        Bytes b3 = new Bytes("5gb");
+        Bytes b4 = new Bytes("5120 b");
+
+        assertTrue(b1.equals(b4));
+        assertFalse(b2.equals(null));
+        assertTrue(b2.equals(b3));
+        assertFalse(b1.equals(b2));
+    }
+
+    @Test
+    public void testHashCode() {
+        Bytes b1 = new Bytes("5kb");
+        Bytes b2 = new Bytes("5368709120b");
+        Bytes b3 = new Bytes("5gb");
+        Bytes b4 = new Bytes("5120 b");
+
+        assertEquals(b1.hashCode(), b4.hashCode());
+        assertEquals(b2.hashCode(), b3.hashCode());
+        assertNotEquals(b1.hashCode(), b3.hashCode());
+    }
+
+    @Test
+    public void testUnitConversion() {
+        assertEquals(5120, new Bytes("5kb").valueOfUnit(BYTES));
+        assertEquals(5, new Bytes("5kb").valueOfUnit(KB));
+        assertEquals(5, new Bytes("5mb").valueOfUnit(MB));
+        assertEquals(5, new Bytes("5120mb").valueOfUnit(GB));
+    }
+}